Working (somehow)

+++ lab/elipse.c
@@ -1,7 +1,7 @@
 #include <stdlib.h>
 #include "elipse.h"
 
-Elipse* crear_elipse(int o_x, int o_y, double alpha, double theta, int beta_i) {
+Elipse* crear_elipse(double o_x, double o_y, double alpha, double theta, int beta_i) {
   Elipse* elipse = (Elipse*) malloc(sizeof(Elipse));
   elipse->o_x = o_x;
   elipse->o_y = o_y;
