+++ lab/elipse.h
@@ -1,13 +1,13 @@
 #ifndef ELIPSE_H
 #define ELIPSE_H
 typedef struct elipse {
-  int o_x;
-  int o_y;
+  double o_x;
+  double o_y;
   double alpha;
   double theta;
   int beta_i;
 }Elipse;
 
-Elipse* crear_elipse(int o_x, int o_y, double alpha, double theta, int beta_i);
+Elipse* crear_elipse(double o_x, double o_y, double alpha, double theta, int beta_i);
 
 #endif /* ELIPSE_H */
